requirements.txt
pandas>=2.0
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
celery>=5.3
redis>=5.0
python-dotenv
structlog
click
PyYAML>=6.0
psycopg2-binary>=2.9
